[Feat]: settings 추가 수정
- api문서 수정에 따른 경로 및 함수 이름 수정
